Update README.md
Upon reexamination, I realized that the shader.frag and shader.vert were not mine.  To clarify, I originally wrote a camera class that utilized two files named shader.frag and shader.vert.  That class was not used in the final submission of the project and it would appear that one of the other team members rewrote shader.frag and shader.vert.  Therefore I am updating the README file so that I am no longer taking credit for those files.